Refactor: MainActivity 포그라운드 서비스 시작 로직 삭제
